+++ messages/javascript/src/index.ts
@@ -1,20 +1,7 @@
-import MessageToBinaryStream from './MessageToBinaryStream'
-import MessageToNdjsonStream from './MessageToNdjsonStream'
-import BinaryToMessageStream from './BinaryToMessageStream'
-import NdjsonToMessageStream from './NdjsonToMessageStream'
 import * as TimeConversion from './TimeConversion'
 import * as IdGenerator from './IdGenerator'
 import { io } from './messages'
 import messages = io.cucumber.messages
 import { version } from '../package.json'
 
-export {
-  messages,
-  MessageToBinaryStream,
-  MessageToNdjsonStream,
-  BinaryToMessageStream,
-  NdjsonToMessageStream,
-  TimeConversion,
-  IdGenerator,
-  version,
-}
+export { messages, TimeConversion, IdGenerator, version }